refactoring to simplify code analysis
